UI revamp and status control polishing
UI remodeled (still needs some work as difficulty and error margin are
not editable yet). The interface is still very primitive, using text
boxes.

Polishing of the status flow scheme to make it more robust and clear.

The priority right now is implementing a input box for difficulty and
error margin.