Added rep. counter/modifier
Added a repetition counter and a text box to modify the internal value.
In addition to that, added a method to reset the parameters between
repetititons. The menu routine now resets to 0 the repetition count. The
first routine has been modified to adapt it to this new menu-model.
Still needs more user interaction.